Initial commit of puppet-sssd module
This is a simple module to configure SSSD for RedHat systems. The
objective is to make this as lightweight as possible. As there are a
very large number of configuration items, support a free-form
configuration options hash. This has takes the form of:
<section> => {
    <config_option> => <config_option_value>,
    ...
    <config_option> => <config_option_value>,
}

Configuring SSSD in this way will ensure that this module is always
capable of configuring SSSD despite new or changed config options.